delete endpoint added get all users endpoint added exceptions added

+++ src/main/java/se/frisk/cadettsplittersgateway_edufy/services/KeycloakServiceImpl.java
@@ -5,6 +5,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import se.frisk.cadettsplittersgateway_edufy.clients.KeycloakClient;
 import se.frisk.cadettsplittersgateway_edufy.dtos.KeycloakDTO;
+import se.frisk.cadettsplittersgateway_edufy.dtos.UserRepresentation;
+import se.frisk.cadettsplittersgateway_edufy.exceptions.UserAlreadyExistsException;
+import se.frisk.cadettsplittersgateway_edufy.exceptions.UserNotFoundException;
+import se.frisk.cadettsplittersgateway_edufy.utils.DTOConvertor;
+
+import java.util.List;
 
 @Service
 public class KeycloakServiceImpl implements KeycloakService {
@@ -27,25 +33,34 @@ public class KeycloakServiceImpl implements KeycloakService {
 
         String existingUserId = getKeycloakUserId(userDto.getUsername());
         if (existingUserId != null) {
-            System.out.println("User '" + userDto.getUsername() + "' already exists with ID: " + existingUserId);
-            return existingUserId;
+            throw new UserAlreadyExistsException("user", "keycloak id", existingUserId);
         }
 
         String existingUserEmail = userDto.getEmail();
-        /// add method to check the email is not already in use
+        if (keycloakClient.emailExists(existingUserEmail)){
+            throw new UserAlreadyExistsException("user", "email", existingUserEmail);
+        }
 
 
         return keycloakClient.createKeycloakUser(userDto);
     }
 
+
+
     @Override
     public String getKeycloakUserId(String username) {
-        return keycloakClient.getKeycloakUserId(username);
+        String userId = keycloakClient.getKeycloakUserId(username);
+        if (userId == null) {
+            throw new UserNotFoundException(username);
+        } return userId;
     }
 
 
     @Override
     public void deleteKeycloakUser(String keycloakId) {
+        if(keycloakClient.getUserById(keycloakId) == null) {
+            throw new UserNotFoundException(keycloakId);
+        }
         keycloakClient.deleteKeycloakUser(keycloakId);
     }
 
@@ -56,12 +71,24 @@ public class KeycloakServiceImpl implements KeycloakService {
         if (userId == null) {
             userId = keycloakClient.getKeycloakUserId(userDto.getUsername());
             if (userId == null) {
-                throw new RuntimeException("User not found: " + userDto.getUsername());
+                throw new UserNotFoundException(userDto.getUsername());
             }
             userDto.setKeycloakId(userId);
         }
 
         keycloakClient.assignRoleToUser(userDto);
+
+    }
+
+    @Override
+    public List<KeycloakDTO> getAllKeycloakUsers() {
+        List<UserRepresentation> users = keycloakClient.getAllUsers();
+
+        List<KeycloakDTO> dtos = users.stream()
+                .map(DTOConvertor::toKeycloakDTO)
+                .toList();
+
+        return dtos;
 
     }
